requirements.txt
python-dateutil
discord.py
iteration_utilities
